refactor: group.committees.add! and group.add_council!
The old way we did this was jumping through quite a few hoops because when creating the Committee (or Council) we still did not set the parent. So the parent was only set in a second step and only then permissions could be set properly.

In addition to that i am striving towards a new more standardized API on top of AR.

I don't want to mess with the AR functions like create or destroy and their callbacks for caching and the like anymore. It's messy to turn off callbacks when you don't need them and hard to tell which callbacks will be triggered.

So instead i am starting to standartize high level api calls
that build upon AR and also handle our special caching stuff etc.:

* group.committees.add! - create a committee and add it to the group.
* group.add_council!    - create the council for the group

In general i want to mimic the rails convention of

owner.collection.verb   - act upon a collection (has_many)
owner.verb_target       - act upon a single target (has_one, belongs_to)

And i want to use two verbs to start with:

* add    - create an association and handle the caching etc.
* remove - delete an association and handle caching

So I imagine to also apply this in other places like
group.users.add!
or
page.add_owner!

In some places we hide the underlying low level API. While i don't want
to use it from the controllers I still want to expose it for
maintainance work.
For example protecting #destroy breaks #destroy_all and dependent:
:destroy callbacks. So instead I hope to make it obvious which function
to call by relying on the above convention.

This is not meant to land in crabgrass before version 0.7.0