upgrade: remove need for joining users table
The third column in the order clause was referencing the users table
for sorted user_participations. So basically this means participations#
that were accessed at the same time and changed at the same time would
be sorted by the user login... pretty pointless.

But at the same time it made queries way more complex as the users table
had to be joined.

I noticed this because since we are not autojoining included tables in
rails4 anymore the users join was missing and a .references(:user) would
have been needed. Instead of adding this i removed the last order clause
for sake of simplicity.