upgrade: test with exceptions_app always
So far we set consider_all_requests_local to false in a test setup
for integration tests.
This did not always work on 'rake test' runs. No idea why.

So instead we now have
  show_exception = true
  consider_all_requests_local = false
in the whole test environment.

This way the exceptions app should always be used and render our nice
404, etc. pages